refactor: merge view/ into Views/, update BiLineItemView namespace to Views
