+++ src/Ksfraser/FaBankImport/controller/BiLineItemController.php
@@ -2,7 +2,7 @@
 <namespace Ksfraser\FaBankImport\Controller;
 
 use Ksfraser\FaBankImport\Model\BiLineItemModel;
-use Ksfraser\FaBankImport\View\BiLineItemView;
+use Ksfraser\FaBankImport\Views\BiLineItemView;
 
 /**
  * Controller class for managing line item interactions.
